+++ fltk/src/menu.rs
@@ -343,6 +343,40 @@ impl MenuItem {
     pub fn was_deleted(&self) -> bool {
         self._inner.is_null()
     }
+
+    /// Draw a box around the menu item.
+    /// Requires the call to be made inside a MenuExt-implementing widget's own draw method
+    pub fn draw<M: MenuExt>(&self, x: i32, y: i32, w: i32, h: i32, menu: &M, selected: bool) {
+        assert!(!self.was_deleted());
+        unsafe {
+            Fl_Menu_Item_draw(
+                self._inner,
+                x,
+                y,
+                w,
+                h,
+                menu.as_widget_ptr() as _,
+                selected as i32,
+            )
+        }
+    }
+
+    /// Measure the width and height of a menu item
+    pub fn measure(&self) -> (i32, i32) {
+        assert!(!self.was_deleted());
+        let h: &mut i32 = &mut 0;
+        let ret = unsafe { Fl_Menu_Item_measure(self._inner, h as _, std::ptr::null()) };
+        (ret, *h)
+    }
+
+    /// Set the image of the menu item
+    /// # Safety
+    /// Trying to add a label after adding an image might lead to undefined behavior
+    pub unsafe fn set_image<I: ImageExt>(&mut self, mut image: I) {
+        assert!(!self.was_deleted());
+        image.increment_arc();
+        Fl_Menu_Item_image(self._inner, image.as_image_ptr() as _)
+    }
 }
 
 /// Delete a menu item